Adds certmagic test case and panics recovery
Adds a test case to check main execution path with
certmagic enabled. This test checks HTTPS server initialization
using a dummy domain to ensure that it will not block execution
in test environments.

Adds a recover statement within the test goroutines to prevent panics
from crashing the test suite and to log the error.